# public/docs/admin/swagger.json
{
	"openapi": "3.0.0",
	"components": {
...
		"schemas": {
...
			"Error": {
				"description": "Represents an error in the API response.",
				"properties": {
					"code": {
						"type": "string",
						"description": "The error code."
					},
					"message": {
						"type": "string",
						"description": "The error message."
					}
				},
				"required": [
					"code",
					"message"
				],
				"type": "object",
				"additionalProperties": false
			},
			"Warning": {
				"description": "Represents a warning in the API response.",
				"properties": {
					"code": {
						"type": "string",
						"description": "The warning code."
					},
					"message": {
						"type": "string",
						"description": "The warning message."
					}
				},
				"required": [
					"code",
					"message"
				],
				"type": "object",
				"additionalProperties": false
			},
...
			"CountryInfo": {
				"description": "Represents information about a country.",
				"properties": {
					"iso2": {
						"type": "string",
						"description": "The ISO 3166-1 alpha-2 code associated with the country."
					},
					"iso3": {
						"type": "string",
						"description": "The ISO 3166-1 alpha-3 code associated with the country."
					},
					"name": {
						"type": "string",
						"description": "The name of the country."
					}
				},
				"required": [
					"iso2",
					"iso3",
					"name"
				],
				"type": "object",
				"additionalProperties": false,
				"example": {
					"iso2": "US",
					"iso3": "USA",
					"name": "United States"
				}
			},
...
			"StateInfo": {
				"description": "Represents information about a state.",
				"properties": {
					"id": {
						"type": "number",
						"format": "double",
						"description": "The unique identifier (ID) of the state."
					},
					"name": {
						"type": "string",
						"description": "The name of the state."
					}
				},
				"required": [
					"id",
					"name"
				],
				"type": "object",
				"additionalProperties": false
			},
			"CityInfo": {
				"description": "Represents information about a city.",
				"properties": {
					"id": {
						"type": "number",
						"format": "double",
						"description": "The unique identifier (ID) of the city."
					},
					"name": {
						"type": "string",
						"description": "The name of the city."
					}
				},
				"required": [
					"id",
					"name"
				],
				"type": "object",
				"additionalProperties": false
			},
...
			"LanguageInfo": {
				"description": "Represents information about a language.",
				"properties": {
					"code": {
						"type": "string",
						"description": "The code associated with the language."
					},
					"name": {
						"type": "string",
						"description": "The name of the language."
					}
				},
				"required": [
					"code",
					"name"
				],
				"type": "object",
				"additionalProperties": false,
				"example": {
					"code": "en",
					"name": "English"
				}
			},
			"LanguagesData": {
				"description": "Represents the data structure for languages.\nExtends List with a generic parameter of LanguageInfo type.",
				"properties": {
					"items": {
						"items": {
							"$ref": "#/components/schemas/LanguageInfo"
						},
						"type": "array",
						"description": "An array of items."
					}
				},
				"required": [
					"items"
				],
				"type": "object",
				"additionalProperties": false
			},
			"GetLanguagesRes": {
				"description": "Represents the response payload for retrieving a list of languages.\nExtends the base ApiResponse class with a generic parameter of LanguagesData type.",
				"properties": {
					"data": {
						"allOf": [
							{
								"$ref": "#/components/schemas/LanguagesData"
							}
						],
						"nullable": true,
						"description": "The data in the response, if any."
					},
					"errors": {
						"items": {
							"$ref": "#/components/schemas/Error"
						},
						"type": "array",
						"description": "An array of errors in the response."
					},
					"warnings": {
						"items": {
							"$ref": "#/components/schemas/Warning"
						},
						"type": "array",
						"description": "An array of warnings in the response."
					}
				},
				"type": "object",
				"additionalProperties": false
			},
			"CountriesData": {
				"description": "Represents the data structure for countries.\nExtends List with a generic parameter of CountryInfo type.",
				"properties": {
					"items": {
						"items": {
							"$ref": "#/components/schemas/CountryInfo"
						},
						"type": "array",
						"description": "An array of items."
					}
				},
				"required": [
					"items"
				],
				"type": "object",
				"additionalProperties": false
			},
			"GetCountriesRes": {
				"description": "Represents the response payload for retrieving a list of countries.\nExtends the base ApiResponse class with a generic parameter of CountriesData type.",
				"properties": {
					"data": {
						"allOf": [
							{
								"$ref": "#/components/schemas/CountriesData"
							}
						],
						"nullable": true,
						"description": "The data in the response, if any."
					},
					"errors": {
						"items": {
							"$ref": "#/components/schemas/Error"
						},
						"type": "array",
						"description": "An array of errors in the response."
					},
					"warnings": {
						"items": {
							"$ref": "#/components/schemas/Warning"
						},
						"type": "array",
						"description": "An array of warnings in the response."
					}
				},
				"type": "object",
				"additionalProperties": false
			},
			"StateData": {
				"description": "Represents information about a state.",
				"properties": {
					"state": {
						"$ref": "#/components/schemas/StateInfo",
						"description": "Information about the state."
					},
					"cities": {
						"items": {
							"$ref": "#/components/schemas/CityInfo"
						},
						"type": "array",
						"description": "List of cities within the state."
					}
				},
				"required": [
					"state",
					"cities"
				],
				"type": "object",
				"additionalProperties": false
			},
			"CountryData": {
				"description": "Represents detailed information about a country.",
				"properties": {
					"country": {
						"$ref": "#/components/schemas/CountryInfo",
						"description": "Information about the country."
					},
					"states": {
						"items": {
							"$ref": "#/components/schemas/StateData"
						},
						"type": "array",
						"description": "List of states within the country."
					}
				},
				"required": [
					"country",
					"states"
				],
				"type": "object",
				"additionalProperties": false,
				"example": {
					"country": {
						"iso2": "US",
						"iso3": "USA",
						"name": "United States"
					},
					"states": [
						{
							"state": {
								"id": 1,
								"name": "California"
							},
							"cities": [
								{
									"id": 123,
									"name": "Los Angeles"
								}
							]
						}
					]
				}
			},
			"GetCountryRes": {
				"description": "Represents the response payload for retrieving information about a country.\nExtends the base ApiResponse class with a generic parameter of CountryData type.",
				"properties": {
					"data": {
						"allOf": [
							{
								"$ref": "#/components/schemas/CountryData"
							}
						],
						"nullable": true,
						"description": "The data in the response, if any."
					},
					"errors": {
						"items": {
							"$ref": "#/components/schemas/Error"
						},
						"type": "array",
						"description": "An array of errors in the response."
					},
					"warnings": {
						"items": {
							"$ref": "#/components/schemas/Warning"
						},
						"type": "array",
						"description": "An array of warnings in the response."
					}
				},
				"type": "object",
				"additionalProperties": false
			},
...
		},
		"securitySchemes": {
			"AccessToken": {
				"type": "apiKey",
				"name": "x-access-token",
				"in": "header",
				"scopes": {
					"admin": "Admin"
				}
			},
			"RefreshToken": {
				"type": "apiKey",
				"name": "x-refresh-token",
				"in": "header",
				"scopes": {
					"admin": "Admin"
				}
			}
		}
	},
	"info": {
		"title": "Node.js tsoa MongoDB Boilerplate Admin API",
		"version": "v1",
		"description": "This is a Node.js tsoa MongoDB Boilerplate Admin API version v1. It provides comprehensive documentation and administrative functionalities for managing MongoDB databases. The API is designed to streamline administrative tasks and enhance the overall efficiency of MongoDB operations. The 'public/docs/admin' directory serves as the output location for the API documentation, ensuring easy access for developers and administrators. With specVersion 3, this API adheres to the latest specifications, offering robust performance and compatibility. Whether it's managing user permissions, data schemas, or database configurations, this API simplifies administrative workflows and empowers developers to build scalable MongoDB applications with ease.",
		"license": {
			"name": "Apache 2.0"
		},
		"contact": {
			"name": "Akram B.",
			"email": "belhajsalem.akram@gmail.com"
		}
	},
	"paths": {
...
		"/directories/languages": {
			"get": {
				"operationId": "getLanguages",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/GetLanguagesRes"
								}
							}
						}
					}
				},
				"description": "Get supported languages.",
				"summary": "Get a list of supported languages.",
				"tags": [
					"Multilingual Directories"
				],
				"security": [
					{
						"AccessToken": [
							"admin"
						]
					},
					{
						"RefreshToken": [
							"admin"
						]
					}
				],
				"parameters": [
					{
						"description": "- The language code.",
						"in": "header",
						"name": "x-lang-code",
						"required": false,
						"schema": {
							"default": "en",
							"type": "string"
						}
					}
				]
			}
		},
		"/directories/countries": {
			"get": {
				"operationId": "getCountries",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/GetCountriesRes"
								}
							}
						}
					}
				},
				"description": "Get supported countries.",
				"summary": "Get a list of supported countries.",
				"tags": [
					"Multilingual Directories"
				],
				"security": [
					{
						"AccessToken": [
							"admin"
						]
					},
					{
						"RefreshToken": [
							"admin"
						]
					}
				],
				"parameters": [
					{
						"description": "- The language code.",
						"in": "header",
						"name": "x-lang-code",
						"required": false,
						"schema": {
							"default": "en",
							"type": "string"
						}
					}
				]
			}
		},
		"/directories/countries/{iso3}": {
			"get": {
				"operationId": "getCountry",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/GetCountryRes"
								}
							}
						}
					}
				},
				"description": "Get information about a specific country.",
				"summary": "Get information about a specific country by ISO3 code.",
				"tags": [
					"Multilingual Directories"
				],
				"security": [
					{
						"AccessToken": [
							"admin"
						]
					},
					{
						"RefreshToken": [
							"admin"
						]
					}
				],
				"parameters": [
					{
						"description": "- The language code.",
						"in": "header",
						"name": "x-lang-code",
						"required": false,
						"schema": {
							"default": "en",
							"type": "string"
						}
					},
					{
						"description": "- The ISO3 code of the country to retrieve information about.",
						"in": "path",
						"name": "iso3",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
...
	},
	"servers": [
		{
			"url": "/api/admin/v1"
		}
	],
...
}